requirements.txt
selenium~=4.7.2
Jinja2~=3.1.2
pytest~=7.2.1
webdriver-manager~=3.8.2
pytest~=7.2.1
allure-pytest~=2.12.0
HTMLTestReport~=1.0.1
pip~=22.2.2
requests~=2.28.2